creating the login page without testing
