+++ frontend/src/components/login/Login.tsx
@@ -1,56 +1,19 @@
-import React, { FormEvent, useState, useContext } from "react";
+import React, { useContext } from "react";
 import { LoginContext } from "../login/LoginManager";
 import { Redirect } from "react-router-dom";
+import { LoginForm } from "./LoginForm";
 
 export const Login: React.FunctionComponent = () => {
-  const [username, setUsername] = useState("");
-  const [password, setPassword] = useState("");
-  const [error, setError] = useState<string>();
   const loginContext = useContext(LoginContext);
-
-  function tryLogin(event: FormEvent) {
-    event.preventDefault();
-    loginContext.logIn(username, password).then((didLogin) => {
-      if (didLogin) {
-        setError(undefined);
-      } else {
-        setError("Login failed.");
-      }
-    });
-  }
 
   if (loginContext.isLoggedIn) {
     return <Redirect to="/" />;
   }
 
   return (
-    <div>
-      {error && <p>{error}</p>}
-      <h1>Log In</h1>
-      <form onSubmit={tryLogin}>
-        <label className="field">
-          Username
-          <input
-            type={"text"}
-            value={username}
-            placeholder="Username"
-            required
-            onChange={(event) => setUsername(event.target.value)}
-          />
-        </label>
-        <label className="field">
-          Password
-          <input
-            type={"password"}
-            value={password}
-            placeholder="Password"
-            required
-            onChange={(event) => setPassword(event.target.value)}
-          />
-        </label>
-
-        <button type="submit">Log In</button>
-      </form>
-    </div>
+    <>
+      <h1>Login Form</h1>
+      <LoginForm />
+    </>
   );
 };
